+++ src/ts/CmcdIntegration.ts
@@ -19,7 +19,7 @@ import {
   SegmentInfo,
   VideoAdaptationData,
 } from 'bitmovin-player';
-import type { CmcdBase } from './Cmcd';
+import { CmcdBase, CmcdCustomKey } from './Cmcd';
 import {
   CmcdBufferLength,
   CmcdBufferStarvation,
@@ -48,16 +48,28 @@ import {
   CmcdVersionNumbers,
 } from './Cmcd';
 
+export { CmcdCustomKey as CustomKey } from './Cmcd';
+
 export interface CmcdConfig {
-  sessionId: string;
-  contentId: string;
+  sessionId?: string;
+  contentId?: string;
   useQueryArgs?: boolean;
+  /**
+   * Allows to set custom keys, which will be appended to all requests.
+   * This list is static and cannot be changed during run-time.
+   *
+   * Please note that custom key names MUST carry a hyphenated prefix to ensure
+   * that there won't be a namespace collision with future revisions of CMCD.
+   * The prefix SHOULD also use a reverse-DNS syntax.
+   */
+  customKeys?: CmcdCustomKey[];
 }
 
 export class CmcdIntegration {
   private sessionId: string;
   private contentId: string;
   private useQueryArgs: boolean;
+  private customKeys: CmcdCustomKey[];
   private player?: PlayerAPI;
   private stalledSinceLastRequest: boolean;
   private currentVideoQuality: { bandwidth: number; id: string } | null;
@@ -69,8 +81,9 @@ export class CmcdIntegration {
 
   constructor(config: CmcdConfig) {
     this.useQueryArgs = config.useQueryArgs || false;
-    this.sessionId = config.sessionId;
-    this.contentId = config.contentId;
+    this.sessionId = config.sessionId || '';
+    this.contentId = config.contentId || '';
+    this.customKeys = config.customKeys || [];
     this.stalledSinceLastRequest = false;
     this.currentVideoQuality = null;
     this.currentAudioQuality = null;
@@ -107,9 +120,12 @@ export class CmcdIntegration {
     const data = this.gatherCmcdData(type, request);
 
     if (this.useQueryArgs) {
+      const url = new URL(request.url);
+      url.searchParams.delete('CMCD');
+
       const cmcdStr = cmcdDataToUrlParameter(data);
-      const separator = request.url.includes('?') ? '&' : '?';
-      request.url = `${request.url}${separator}${cmcdStr}`;
+      const separator = url.toString().includes('?') ? '&' : '?';
+      request.url = `${url.toString()}${separator}${cmcdStr}`;
     } else {
       const cmcdHeaders = cmcdDataToHeader(data);
 
@@ -140,6 +156,14 @@ export class CmcdIntegration {
     }
     return Promise.resolve(response);
   };
+
+  public setSessionId(id: string): void {
+    this.sessionId = id;
+  }
+
+  public setContentId(id: string): void {
+    this.contentId = id;
+  }
 
   private gatherCmcdData(type: HttpRequestType, request: HttpRequest): CmcdBase[] {
     if (!this.player) {
@@ -179,6 +203,10 @@ export class CmcdIntegration {
     data.concat(this.getRequestedMaximumThroughput(data));
 
     // TODO: data.push(new CmcdNextRangeRequest('byte-range'));
+
+    for (const customKey of this.customKeys) {
+      data.push(customKey);
+    }
 
     return data;
   }
@@ -290,12 +318,16 @@ export class CmcdIntegration {
     }
 
     if (this.currentVideoQuality) {
-      const allSegments = this.player.getAvailableSegments();
-      for (const mimeType in allSegments) {
-        if (mimeType.startsWith('video/')) {
-          const segments = allSegments[mimeType][this.currentVideoQuality.id];
-          data = data.concat(this.getNextObjectAndObjectDurationCmcdData(segments, request.url));
+      try {
+        const allSegments = this.player.getAvailableSegments();
+        for (const mimeType in allSegments) {
+          if (mimeType.startsWith('video/')) {
+            const segments = allSegments[mimeType][this.currentVideoQuality.id];
+            data = data.concat(this.getNextObjectAndObjectDurationCmcdData(segments, request.url));
+          }
         }
+      } catch (error) {
+        console.error('Error processing audio segments:', error);
       }
     }
 
@@ -337,12 +369,16 @@ export class CmcdIntegration {
     }
 
     if (this.currentVideoQuality) {
-      const allSegments = this.player.getAvailableSegments();
-      for (const mimeType in allSegments) {
-        if (mimeType.startsWith('video/')) {
-          const segments = allSegments[mimeType][this.currentVideoQuality.id];
-          data = data.concat(this.getNextObjectAndObjectDurationCmcdData(segments, request.url));
+      try {
+        const allSegments = this.player.getAvailableSegments();
+        for (const mimeType in allSegments) {
+          if (mimeType.startsWith('video/')) {
+            const segments = allSegments[mimeType][this.currentVideoQuality.id];
+            data = data.concat(this.getNextObjectAndObjectDurationCmcdData(segments, request.url));
+          }
         }
+      } catch (error) {
+        console.error('Error processing video segments:', error);
       }
     }